[TextWrapper] Clean up spacing
